requirements.txt
fastapi==0.95.0
uvicorn==0.23.2
python-dotenv==1.0.0
authlib==1.2.1
jinja2==3.1.2
itsdangerous==2.2.0
httpx